Invalid Word Removal
Added the remove_word_from_json function which removes a given word from my json file that has the dictionary words (i.e. all possible words that can be used). This is useful because my dictionary doesn't completely line up with the acceptable words in the NYT app so this way if there is a word that is suggested in a solution that isn't allowed to be input I can permanently remove it from the dictionary, so that it never shows up in a possible solution again.