Bug Fixes
Length of letters threshhold was at 16 when it should've been 12 (this was because I originally thought there were 4 letters per side and not 3). I also allowed for the possibility of the first word being a word that didn't have the best score if the best score words were shown to not lead to a solution.